Adjusting corrector
Leave cache management to it and manage data save.

+++ corrector.py
@@ -6,23 +6,23 @@ import multiprocessing as mp
 import numpy as np
 import glob
 import json
-import shutil
 import configparser
+import socket
 
 class Corrector:
     def __init__(
         self,
-        nbins,
-        steps = None,
-        iters = None,
-        nsamples = None,
-        folderName = None,
-        cacheDir = None,
-        workers=1,
-        smoothed=False,
-        display=False,
-        retrieve=True,
-        config = None,
+        nbins:int,
+        steps:int = None,
+        iters:int = None,
+        nsamples:int = None,
+        folderName:str = None,
+        cacheDir:str = None,
+        workers:int=1,
+        smoothed:bool=False,
+        display:bool=False,
+        retrieve:bool=True,
+        config:str|configparser.ConfigParser = None,
         **kwargs
     ):
         self.nbins = nbins
@@ -45,6 +45,20 @@ class Corrector:
             self.steps = config.getint("correction", "steps", fallback=200)
             self.iters = config.getint("correction", "iters", fallback=10000)
             self.nsamples = config.getint("correction", "nsamples", fallback=0)
+            self.cacheDir = self.config.getint("correction", "cacheDir", fallback=None)
+
+            thisHost = socket.gethostname()
+            if self.config.has_section(thisHost):
+                self.cacheDir = self.config.get(thisHost, "cacheDir", fallback=self.cacheDir)
+
+            path = os.path.dirname(os.path.realpath(__file__))
+            if self.cacheDir is not None and not os.path.isdir(self.cacheDir) and os.path.isdir(os.path.join(path, self.cacheDir)):
+                self.cacheDir = os.path.join(path, self.cacheDir)
+
+        if cacheDir is not None:
+            self.cacheDir = cacheDir
+            if self.cacheDir is not None and not os.path.isdir(self.cacheDir) and os.path.isdir(os.path.join(path, self.cacheDir)):
+                self.cacheDir = os.path.join(path, self.cacheDir)
         
         if steps is not None:
             self.steps = steps
@@ -54,7 +68,6 @@ class Corrector:
             self.nsamples = nsamples
 
         self.folderName = folderName
-        self.cacheDir = cacheDir
 
         self.smoothed = smoothed
         self.display = display
@@ -64,13 +77,23 @@ class Corrector:
         self.trueval = None
         self.old_correctI = np.vectorize(self._correctI)
         self.correctI = lambda x: correct_vector(x, self.newco, self.trueval)
-        if retrieve:
-            self.__retrieve()
+        
+        self.__retrieve(retrieve)
     
-    def __retrieve (self):
-        if os.path.isfile(os.path.join(self.folderName,f"newco_{self.nbins}.npy")) or os.path.isfile(os.path.join(self.cacheDir,f"newco_{self.nbins}_{self.nsamples}.npy")):
+    def __retrieve (self, retrieve):
+        self.earlyResultsPath = None
+
+        if os.path.isfile(os.path.join(self.folderName,f"newco_{self.nbins}.npy")):
+            self.earlyResultsPath = os.path.join(self.folderName,f"newco_{self.nbins}.npy")
             return
         
+        if os.path.isfile(os.path.join(self.cacheDir,f"newco_{self.nbins}_{self.nsamples}.npy")):
+            self.earlyResultsPath = os.path.join(self.cacheDir,f"newco_{self.nbins}_{self.nsamples}.npy")
+            return
+        
+        if not retrieve:
+            return
+            
         for fold in glob.glob(os.path.abspath(os.path.join(self.folderName, os.pardir, f"*bin{self.nbins}"))):
             if os.path.isfile(os.path.join(fold,"shape.json")):
                 with open(os.path.join(fold,"shape.json")) as fp:
@@ -78,18 +101,14 @@ class Corrector:
                     if shape[0] == self.nsamples:
                         if os.path.isfile(os.path.join(fold,f"newco_{self.nbins}.npy")):
                             print("Retrieving correction from: ", fold)
-                            shutil.copy2(os.path.join(fold,f"newco_{self.nbins}.npy"), self.folderName)
-                            if os.path.isfile(os.path.join(fold,f"trueval_{self.nbins}.npy")):
-                                shutil.copy2(os.path.join(fold,f"trueval_{self.nbins}.npy"), self.folderName)
-                            break
+                            self.earlyResultsPath = os.path.join(fold,f"newco_{self.nbins}.npy")
+                            return
 
     def compute_correction(self):
         """Computes the correction lookup table or loads the cached values."""
-        / cache e salvataggio   
         incre = 1 / self.steps
         correction = np.zeros(self.steps)
-        self.trueval = np.zeros(self.steps)
-        if not os.path.isfile(f"{self.folderName}/newco_{self.nbins}.npy"):
+        if self.earlyResultsPath is None:
             pool = mp.Pool(self.workers)
             for i in tqdm(range(self.steps), "Computing correction"):
                 means = 0, 0
@@ -102,7 +121,6 @@ class Corrector:
                     ),
                 )
                 correction[i] = np.average(I)
-                self.trueval[i] = -0.5 * np.log(1 - (i * incre) ** 2)
 
             if self.smoothed:
                 tosmo = np.array(
@@ -128,52 +146,44 @@ class Corrector:
                         pass
             else:
                 self.newco = correction
-
-            np.save(f"{self.folderName}/newco_{self.nbins}.npy", self.newco)
-            np.save(f"{self.folderName}/trueval_{self.nbins}.npy", self.trueval)
             pool.close()
         else:
-            self.newco = np.load(f"{self.folderName}/newco_{self.nbins}.npy")
+            self.newco = np.load(self.earlyResultsPath)
             correction = self.newco
-            if os.path.isfile(f"{self.folderName}/trueval_{self.nbins}.npy"):
-                self.trueval = np.load(f"{self.folderName}/trueval_{self.nbins}.npy")
-            else:
-                self.trueval = -0.5 * np.log(
-                    1 - (np.arange(self.steps) / self.steps) ** 2
-                )
-                np.save(f"{self.folderName}/trueval_{self.nbins}.npy", self.trueval)
 
+        self.trueval = -0.5 * np.log(1 - (np.arange(self.steps) / self.steps) ** 2)
+
+        if self.cacheDir and not self.cacheDir in self.earlyResultsPath:
+            np.save(os.path.join(self.cacheDir, f"newco_{self.nbins}_{self.nsamples}.npy", self.newco))
+
+        if self.folderName and not self.folderName in self.earlyResultsPath:
+            np.save(os.path.join(self.folderName, f"newco_{self.nbins}.npy", self.newco))
+
+        if self.folderName or self.display:
+            # this is needed to get an estimate of the size of the bias we are correcting
             weights = np.zeros_like(self.trueval)
             weights[:-1] += 0.5 * (self.trueval[1:] - self.trueval[:-1])
             weights[1:] += weights[:-1]
             deviation = np.sqrt(
-                np.average(np.square(self.newco[:] - self.trueval[:]), weights=weights)
+                np.average(np.square(correction[:] - self.trueval[:]), weights=weights)
             )
 
-        # this is needed to get an estimate of the size of the bias we are correcting
-        weights = np.zeros_like(self.trueval)
-        weights[:-1] += 0.5 * (self.trueval[1:] - self.trueval[:-1])
-        weights[1:] += weights[:-1]
-        deviation = np.sqrt(
-            np.average(np.square(correction[:] - self.trueval[:]), weights=weights)
-        )
-
-        plt.title(f"RMS correction: {deviation:.4}")
-        plt.plot(self.trueval, correction)
-        plt.plot(self.trueval, self.newco)
-        plt.plot(
-            [min(self.trueval), max(self.trueval)],
-            [min(self.trueval), max(self.trueval)],
-            ":k",
-        )
-        plt.xlabel("True MI")
-        plt.ylabel("Estimated MI")
-        if not os.path.isfile(f"{self.folderName}/correctionMap_{self.nbins}.pdf"):
-            plt.savefig(f"{self.folderName}/correctionMap_{self.nbins}.pdf", bbox_inches="tight")
-        if self.display:
-            plt.show()
-        else:
-            plt.close()
+            plt.title(f"RMS correction: {deviation:.4}")
+            plt.plot(self.trueval, correction)
+            plt.plot(self.trueval, self.newco)
+            plt.plot(
+                [min(self.trueval), max(self.trueval)],
+                [min(self.trueval), max(self.trueval)],
+                ":k",
+            )
+            plt.xlabel("True MI")
+            plt.ylabel("Estimated MI")
+            if self.folderName and not os.path.isfile(f"{self.folderName}/correctionMap_{self.nbins}.pdf"):
+                plt.savefig(f"{self.folderName}/correctionMap_{self.nbins}.pdf", bbox_inches="tight")
+            if self.display:
+                plt.show()
+            else:
+                plt.close()
 
     def _correctI(self, I):
         ind = np.argmin(np.abs(I - self.newco))
